InnoDB compatibilty
there is no obvious reason, except for integrity, to use InnoDB instead
of MyISAM for patstat. The better reason is in case of complex queries
on multiple tables, or if you make heavy use of subqueries, in which case
InnoDB outperforms MyISAM.

The script assumes to use barracuda file-per-table, compressed tables.
If you are trying to use it in mysql < 5.6, consider to use MyISAM
instead.

the flag used to change engine is -e. Default engine is MyISAM.