refactor: tls224 and tls225 (patstat2020a)
Due to a change practise of the EPO and USPTO, CPC classification
symbols are not assigned any more to applications, but to DOCDB
families. The new table TLS225_DOCDB_FAM_CPC reflects this. It
also contains information previously not available.

The existing table TLS224_APPLN_CPC with its most relevant attributes
is redundantly kept for downward compatibility.